Add dependency on DB for storing metadata about compilation
I've chosen xodus db from Jetbrains to store metdata about compilation.
Xodus db is choses as aleranative to SQLLite as I don't want to deal
with SQL. It also has a decent support for kotling with Xodus DNQ. For
now We don't think we need transactional nature of Xodus so the choice
is not really based on the need. But it could be useful in the future
use case if we run compilation in parallel.

As a metadata info we need to store info about the last build
and hash of all files to know whether anything changed or not.
Potentially we could store AST tree if .class files is not enough.